// src/module/carousel/carousel.interface.ts
export interface SlideData {
  slide1Title: string;
  slide1Description: string;
  slide2Title: string;
  slide2Description: string;
  slide3Title: string;
  slide3Description: string;
  slide4Title: string;
  slide4Description: string;
  images: {
    slide1: string;
    slide2: string;
    slide3: string;
    slide4: string;
  };
}